backlog folder structure - getAllTasks

+++ src/components/actions/projectActions.js
@@ -51,14 +51,16 @@ const getProject= (identifier, history) => async dispatch => {
 
 }
 const deleteProject= (identifier) => async dispatch => {
-    try {
-        await axios.delete(`http://localhost:8080/api/project/${identifier}`);
-        dispatch({
-            type: "DELETE_PROJECT",
-            payload: identifier
-        })
-    } catch(error) {
-        console.log(error);
+    if(window.confirm("Are you Sure you want to delete this project?")) {
+        try {
+            await axios.delete(`http://localhost:8080/api/project/${identifier}`);
+            dispatch({
+                type: "DELETE_PROJECT",
+                payload: identifier
+            })
+        } catch(error) {
+            console.log(error);
+        }
     }
 
 }
